serdes_clk_gen: Added synchronization between clocks of same frequency
The module (with the support of an FSM in the upper-level wb_serdes_clk_gen) now
supports synchronization of clocks generated by modules connected to the same
White Rabbit network. The clocks are synchronized to midnight, 01/01/1970 (epoch
reference).

If WR is not present due to no connection, synchronization still applies between
clocks of the same frequency generated by the same module.

This synchronization scheme is selectable via a generic.