serdes_clk_gen: Implemented second bit pattern
The second bit pattern is needed for fractional division rates, as the logic for
inserting a zero in the bit pattern to achieve fractional division rates was
erroneous.

This logic was previously inserting a zero only at the middle of the bit pattern;
with a second bit pattern, the logic can now select between the "normal" bit pattern
when no fractional carry needs to take place (no zero needs to be inserted), or
the "skip" mask, which inserts a zero at the proper place in the bit pattern.

The normal and skip masks still need to be calculated in software and set in the
appropriate registers (NORMMASKR & SKIPMASKR).